Add new file If.cs and related tests
Added a new class file If.cs which contains different tasks involving if statements. This is part of the homework in TasksLesson.
Corresponding test cases for each task have also been added ensuring the correctness of the implemented tasks. The tests check the behavior of the tasks for different inputs, including edge cases.