Update if conditions in If21 and tests
The If21 method has been tweaked to add a new check for when both parameters, x and y, are 0, it now correctly returns 0. This caused updating the corresponding test to verify this new condition.

Also modified the range checks in both condition blocks. The conditions are now less restrictive, allowing the method to return response even when x >= 2 and x >= 3 without having to be less than next integer. Test were adjusted as well this modification.