"Add homework for loops in TasksLesson
Implemented methods for various tasks related to loops in the file `TasksLesson/Homework/For.cs'. These methods were used to replace the Main method of the `TasksLesson`. This was done to complete the homework assigned on loops and to provide greater clarity to the `TasksLesson` structure by separating homework from main lessons."