"Add new Series and While class to the TasksLesson.Homework module"
This commit introduces a new Series class and a While class to the Homework module within the TasksLesson namespace. Each of the classes brings numerous static methods which are intended to cover a wide array of mathematical functions for different use cases. For instance, Series class calculates different types of series while While class includes methods employing while loop for calculations. This is done to expand the module's functionality and provide users with more useful mathematical utilities.