Add MinMax class to Homework module
This commit includes the addition of a new class, MinMax, to the TasksLesson.Homework namespace. This class contains multiple methods to solve and illustrate various 'minmax' scenarios, each presented as a numerical problem. This helps to provide better structure for the homework. The 'Program.cs' file was also updated to switch the execution point to the new MinMax class.