Refactor conditional handling and optimize Series18 method
This commit removes unnecessary 'else' statements which improves code readability in different methods of the Series class. In the case of a '0' detected in several methods, instead of printing the index and breaking the loop, it now only breaks the loop making the actions more consolidated. Furthermore, the method Series18 has been updated, changing the list that stored unique numbers to a hashset for better performance - since checking if a number is part of a hashset is faster than checking a list. Boolean variable 'isAlreadyWritten' in Series17 method has been set to initialise as 'false' which is more logical and intuitive, and the associated conditions have been inversely adapted for this initialization.